fix(backend): type timestamp fields as BigInt across controllers and indexers (#830)
- Update createStream, topUpStreamHandler, withdrawHandler, and sorobanIndexerService to pass BigInt values for startTime, lastUpdateTime, endTime, and timestamp
- Update ClaimableStreamState interface to accept number | bigint for timestamp fields

+++ backend/src/routes/v1/streams/withdraw.ts
@@ -106,7 +106,7 @@ export const withdrawHandler = async (req: AuthenticatedRequest, res: Response) 
       // Call Soroban service
       const result = await sorobanWithdraw(parsedStreamId, req.user.publicKey);
       
-      const now = Math.floor(Date.now() / 1000);
+      const now = BigInt(Math.floor(Date.now() / 1000));
       const nextWithdrawnAmount = (
         BigInt(stream.withdrawnAmount) + BigInt(claimable.claimableAmount)
       ).toString();
